Remove old KittenWebSocketProvider implementation
